Change directory names to underscore
