+++ test/test_app/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "test-app",
+  "name": "test_app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
