+++ test/test_app/src/app/no_implicit_any/no_implicit_any_component.html
