+++ test/test_app/src/app/no_implicit_returns/no_implicit_returns_component.ts
