+++ test/test_app/src/app/strict_property_initialization/strict_property_initialization_component.html
