+++ test/test_app/tsconfig.json
@@ -12,10 +12,10 @@
     "importHelpers": true,
     "target": "es2015",
     "lib": ["es2018", "dom"],
-    "strictNullChecks": true,
-    "noImplicitReturns": true,
-    "noImplicitAny": true,
-    "strictPropertyInitialization": true
+    "strictNullChecks": false,
+    "noImplicitReturns": false,
+    "noImplicitAny": false,
+    "strictPropertyInitialization": false
   },
   "angularCompilerOptions": {
     "fullTemplateTypeCheck": true,
