+++ test/test_files/no_implicit_returns_sample.ts
